STYLE: Update python script to support python 2.x and 3.x
After building against python3, install future:

  ${Slicer_DIR}/../python-install/bin/PythonSlicer -m pip install future

Then run on all files:

  Slicer_DIR=/home/jcfr/Projects/Slicer-Qt5-VTK8-RelWithDebInfo/Slicer-build
  for f in `find ./ -name *.py`; do ${Slicer_DIR}/../python-install/bin/PythonSlicer --launch futurize -w $f; done

+++ Applications/ctkSimplePythonShell/Testing/Python/derivedQWidgetTest.py
@@ -1,3 +1,4 @@
+from __future__ import division
 
 from qt import *
 
@@ -24,7 +25,7 @@ class ExampleWidget(QWidget):
   def center(self):
     screen = QDesktopWidget().screenGeometry()
     size =  self.geometry
-    self.move((screen.width() - size.width())/2, (screen.height() - size.height())/2)
+    self.move(int((screen.width() - size.width())/2), int((screen.height() - size.height())/2))
     
 w = ExampleWidget()
 w.show()
